requirements.txt
nextcord
aiohttp
praw
akinator.py[async]
dismusic
jishaku
nextcord-ext-activities
aiosqlite
